refactor: streamline imports across the application and consolidate domain exports for improved maintainability
