+++ apps/api/rooms/get-room.controller.ts
@@ -1,7 +1,7 @@
 import type { Context } from "hono";
 import type { UserDto } from "../dtos/user.dto.ts";
 import type { buildContainer } from "../middleware/ioc/build-container.ts";
-import { GetRoomQuery } from "@myty/fresh-workspace-domain/rooms/get-room";
+import { GetRoomQuery } from "@myty/fresh-workspace-domain";
 import type { BlankInput } from "hono/types";
 import { HTTPException } from "../http-exception.ts";
