+++ packages/domain/auth-providers/domain.ts
@@ -1,0 +1,1 @@
+export { GetAuthProviderUserCommand } from "./get-auth-provider-user/get-auth-provider-user.command.ts";
