+++ packages/domain/messages/domain.ts
@@ -1,0 +1,2 @@
+export { CreateMessageCommand } from "./create-message/create-message.command.ts";
+export { CreateMessagesSubscriptionByRoomCommand } from "./create-messages-subscription-by-room/create-messages-subscription-by-room.command.ts";
