+++ packages/domain/rooms/domain.ts
@@ -1,0 +1,14 @@
+// Commands
+export { CreateRoomCommand } from "./create-room/create-room.command.ts";
+
+// Queries
+export { GetRoomQuery } from "./get-room/get-room.query.ts";
+export { GetUserRoomsQuery } from "./get-user-rooms/get-user-rooms.query.ts";
+
+// Data Access
+export { type CreateRoomDataAccess } from "./create-room/create-room.command-handler.ts";
+export { type GetUserRoomsDataAccess } from "./get-user-rooms/get-user-rooms.query-handler.ts";
+
+// Responses
+export { type CreateRoomResponse } from "./create-room/create-room.response.ts";
+export { type GetUserRoomsResponse } from "./get-user-rooms/get-user-rooms.response.ts";
